+++ lib/assent.ex
@@ -1,6 +1,22 @@
 defmodule Assent do
   @moduledoc false
 
+  defmodule MissingConfigError do
+    defexception [:key, :config]
+
+    @type t :: %__MODULE__{
+            key: atom(),
+            config: Keyword.t()
+          }
+
+    def message(exception) do
+      key = inspect(exception.key)
+      config_keys = inspect(Keyword.keys(exception.config))
+
+      "Expected #{key} in config, got: #{config_keys}"
+    end
+  end
+
   defmodule CallbackError do
     defexception [:message, :error, :error_uri]
   end
@@ -27,9 +43,9 @@ defmodule Assent do
 
     def message(exception) do
       expected_key = inspect(exception.expected_key)
-      params = inspect(Map.keys(exception.params))
+      param_keys = exception.params |> Map.keys() |> Enum.sort() |> inspect()
 
-      "Expected #{expected_key} in params, got: #{params}"
+      "Expected #{expected_key} in params, got: #{param_keys}"
     end
   end
 
@@ -109,6 +125,49 @@ defmodule Assent do
       Reason:
       #{inspect(exception.reason)}
       """
+    end
+  end
+
+  @doc """
+  Fetches the key value from the configuration.
+
+  Returns a `Assent.MissingConfigError` if the key is not found.
+  """
+  @spec fetch_config(Keyword.t(), atom()) :: {:ok, any()} | {:error, MissingConfigError.t()}
+  def fetch_config(config, key) when is_list(config) and is_atom(key) do
+    case Keyword.fetch(config, key) do
+      {:ok, value} -> {:ok, value}
+      :error -> {:error, MissingConfigError.exception(key: key, config: config)}
+    end
+  end
+
+  @doc """
+  Fetches the key value from the params.
+
+  Returns a `Assent.MissingParamError` if the key is not found.
+  """
+  @spec fetch_param(map(), binary()) :: {:ok, any()} | {:error, MissingParamError.t()}
+  def fetch_param(params, key) when is_map(params) and is_binary(key) do
+    case Map.fetch(params, key) do
+      {:ok, value} -> {:ok, value}
+      :error -> {:error, MissingParamError.exception(expected_key: key, params: params)}
+    end
+  end
+
+  @default_json_library (Code.ensure_loaded?(JSON) && JSON) || Jason
+
+  @doc """
+  Fetches the JSON library in config.
+
+  If not found in provided config, this will attempt to load the JSON library
+  from global application environment for `:assent`. Defaults to
+  `#{inspect(@default_json_library)}`.
+  """
+  @spec json_library(Keyword.t()) :: module()
+  def json_library(config) do
+    case Keyword.fetch(config, :json_library) do
+      :error -> Application.get_env(:assent, :json_library, @default_json_library)
+      {:ok, json_library} -> json_library
     end
   end